==> 01_calculations.Rmd <==
---
title: "Calculations"
date: "`r format(Sys.Date(), '%d %b %Y')`"
output:
  html_document: 
    df_print: paged
    number_sections: no
    toc: yes
    toc_float: true
    toc_depth: 3
    code_folding: show
editor_options:
  chunk_output_type: console
---

```{r setup, echo = FALSE, message=FALSE, warning=FALSE}
# load libraries
library(tidyverse) 
library(readxl)

# source all relevant scripting files
source("scripts/table_functions.R")
source("scripts/growth_functions.R")

# global knitting options
knitr::opts_chunk$set(echo = TRUE)
```

# Experiment data

```{r}
# load experiments and samples
experiments <- read_xlsx("data/experiments.xlsx") %>%
  mutate( 
    # temp and %O2 are always whole integers in this study
    temperature = as.integer(temperature),
    `% O2` = as.integer(`% O2`)
  )
samples <- read_xlsx("data/samples.xlsx") %>%
  mutate(rep = as.integer(parse_number(rep_ID)))
experiments
head(samples)

# cache
experiments %>% write_rds("cache/experiments.rds")
samples %>% write_rds("cache/samples.rds")
```

# Normalized OD

```{r}
# OD multipliers
ministat_OD_scaling <- 3.96 # as per OD measurements culture tube vs. ministat
OD_multipliers <- 
  samples %>%
  left_join(experiments, by = "exp_ID") %>%
  mutate(
    OD_multiplier = if_else(!is.na(ministat) & ministat, 1/ministat_OD_scaling, 1)
  ) %>%
  select(sample_ID, OD_multiplier)

# load growth curve data
growth_data_all <- read_excel("data/growth_data.xlsx") %>%
  left_join(OD_multipliers, by = "sample_ID") %>%
  mutate(
    growth_phase = case_when(
      death_phase ~ "death",
      extended_lag ~ "long lag",
      TRUE ~ "growth"
    ),
    OD_norm = OD_multiplier * OD
  )
head(growth_data_all)

# cache
growth_data_all %>% write_rds("cache/growth_data_all.rds")
```

# Growth rates

```{r}
# calculate growth rates
growth_rates_all <- 
  growth_data_all %>%
  filter(!death_phase, !extended_lag) %>% 
  # fit logistic equation to growth curves
  estimate_growth_curve_parameters(
    time = time.hours,
    N = OD,
    group_by = c(sample_ID)
  ) %>%
  mutate(
    # growth rate in 1/day
    growth_rate.1_day = r * 24, 
    growth_rate_se.1_day = r_se * 24
  )
head(growth_rates_all)

# averages of biological replicates
growth_rates_avg <- 
  growth_rates_all %>%
  left_join(samples, by = "sample_ID") %>%
  group_by(exp_ID) %>%
  summarize(
    n_reps = n(),
    growth_rate_mean.1_day = mean(growth_rate.1_day), 
    growth_rate_sd.1_day = sd(growth_rate.1_day), 
    .groups = "drop"
  )
head(growth_rates_avg)

# cache
growth_rates_all %>% write_rds("cache/growth_rates_all.rds")
growth_rates_avg %>% write_rds("cache/growth_rates_avg.rds")
```

# Generations

```{r}
# extract inoculation OD
inoc_OD <- 
  experiments %>%
  mutate(
    inoc_OD = (str_extract(inoculation, "[0-9.]+\\%") %>% parse_number())/100 * 
      (str_extract(inoculation, "=[0-9.]+") %>% parse_number())
  ) %>%
  select(exp_ID, inoc_OD)

# find final OD
final_OD <- 
  growth_data_all %>%
  group_by(sample_ID) %>%
  summarize(
    OD_end = OD_norm[time.hours == max(time.hours[!death_phase])][1],
    .groups = "drop"
  ) 

generations_all <- 
  final_OD %>%
  left_join(select(samples, sample_ID, exp_ID), by = "sample_ID") %>%
  left_join(inoc_OD, by = "exp_ID") %>% 
  mutate(n_generations = log2( OD_end / inoc_OD )) %>%
  select(-exp_ID)
head(generations_all)

# averages of biological replicates
generations_avg <- 
  generations_all %>%
  left_join(samples, by = "sample_ID") %>%
  group_by(exp_ID) %>%
  summarize(
    n_reps = n(),
    n_generations_mean = mean(n_generations), 
    n_generations_sd = sd(n_generations),
    .groups = "drop"
  ) 
head(generations_avg)

# cache
generations_all %>% write_rds("cache/generations_all.rds")
generations_avg %>% write_rds("cache/generations_avg.rds")
```

# GC data

```{r}
# all GC data
gc_data_all <-
  read_xlsx("data/gc_data.xlsx") %>%
  # factorize compound names
  mutate(
    compound = as_factor(compound) %>%
      fct_recode(
        `14:0` = "C14:0",
        `i15:1` = "i-C15:1 c9", `i15:1` = "i-C15:1 c?", `i15:0` = "i-C15:0", `15:0` = "C15:0",
        `i16:0` = "i-C16:0", `16:1` = "C16:1", `16:0` = "C16:0",
        `i17:1` = "i-C17:1 c9", `i17:0` = "i-C17:0", `17:1` = "C17:1", `17:0` = "C17:0",
        `18:1` = "C18:1 c9", `18:1` = "C18:1 c?", `18:0` = "C18:0",
        `19:1` = "C19:1", `20:0` = "C20:0", 
        `1-i15:0 MAGE` = "1-i-15:0 MAGE", `2-i15:0 MAGE` = "2-i-15:0 MAGE", `1,2-i15:0 DAGE` = "iC15:0 DAGE"
      ) %>%
      fct_relevel(
        "14:0", "i15:1", "i15:0", "15:0", "i16:0", "16:1", "16:0", "i17:1", 
        "i17:0", "17:1", "17:0", "18:1", "18:0", "19:1", "20:0",
        "1-i15:0 MAGE", "2-i15:0 MAGE", "1,2-i15:0 DAGE")
  ) %>%
  # sum isomers
  group_by(sample_ID, compound, compound_class) %>%
  summarize(area = sum(area), amount.ug = sum(amount.ug), .groups = "drop") %>%
  # calculate relative amounts
  group_by(sample_ID) %>%
  mutate(rel_amount = area/sum(area)) %>%
  ungroup() %>%
  # complete dataset
  complete(
    sample_ID, nesting(compound, compound_class),  
    fill = list(area = 0, amount.ug = 0, rel_amount = 0)
  )
head(gc_data_all)

# averages of biological replicates
gc_data_avg <- gc_data_all %>%
  left_join(samples, by = "sample_ID") %>%
  group_by(exp_ID, compound, compound_class) %>%
  summarize(
    n_reps = n(),
    rel_amount_mean = mean(rel_amount),
    rel_amount_sd = sd(rel_amount),
    .groups = "drop"
  )
head(gc_data_avg)

# overall summary
gc_data_sum <- gc_data_avg %>%
  group_by(compound) %>%
  summarize(
    rel_amount_all = mean(rel_amount_mean),
    rel_amount_all_sd = sd(rel_amount_mean),
    .groups = "drop"
  ) %>%
  # normalize
  mutate(rel_amount_all = rel_amount_all / sum(rel_amount_all))
head(gc_data_sum)

# cache
gc_data_all %>% write_rds("cache/gc_data_all.rds")
gc_data_avg %>% write_rds("cache/gc_data_avg.rds")
gc_data_sum %>% write_rds("cache/gc_data_sum.rds")
```

# LC data

```{r}
# standard brGDGTs
standard_brGDGTs <- paste0("brGDGT_", rep(1:3, each = 3), rep(c("a", "b", "c"), times = 3))

# all LC data
lc_data_all <-
  read_xlsx("data/lc_data.xlsx") %>%
  # factorize compound names
  mutate(
    std_brGDGT = compound %in% standard_brGDGTs,
    compound = as_factor(compound) %>%
      fct_recode(
        `brGTGT Ia` = "brGTGT_1a", `brGTGT IIa` = "brGTGT_2a", `brGTGT IIIa` = "brGTGT_3a",
        `brGDGT Ia` = "brGDGT_1a", `brGDGT Ib` = "brGDGT_1b", `brGDGT Ic` = "brGDGT_1c",
        `brGDGT IIa` = "brGDGT_2a", `brGDGT IIb` = "brGDGT_2b", `brGDGT IIc` = "brGDGT_2c",
        `brGDGT IIIa` = "brGDGT_3a", `brGDGT IIIb` = "brGDGT_3b", `brGDGT IIIc` = "brGDGT_3c",
        `brGDGT IIIa-2` = "brGDGT_3a_iso",`brGDGT IIIb-2` = "brGDGT_3b-2"
      ) %>%
      fct_relevel("brGTGT Ia", "brGTGT IIa", "brGTGT IIIa") %>%
      fct_relevel("brGDGT IIIa-2", "brGDGT IIIb-2", after = Inf)
  ) %>%
  # calculate relative amounts
  group_by(sample_ID) %>%
  mutate(
    rel_amount = area/sum(area),
    rel_br = area / sum(area[std_brGDGT])
  ) %>%
  ungroup() %>%
  # complete dataset
  complete(
    sample_ID, nesting(compound, compound_class),  
    fill = list(area = 0, amount.ug = 0, rel_amount = 0, rel_br = 0)
  )
head(lc_data_all)

# averages of biological replicates
lc_data_avg <- lc_data_all %>%
  left_join(samples, by = "sample_ID") %>%
  group_by(exp_ID, compound, std_brGDGT) %>%
  summarize(
    n_reps = n(),
    rel_amount_mean = mean(rel_amount),
    rel_amount_sd = sd(rel_amount),
    rel_br_mean = mean(rel_br),
    rel_br_sd = sd(rel_br),
    .groups = "drop"
  )
head(lc_data_avg)

# overall summary
lc_data_sum <- lc_data_avg %>%
  group_by(compound, std_brGDGT) %>%
  summarize(
    rel_amount_all = mean(rel_amount_mean),
    rel_amount_all_sd = sd(rel_amount_mean),
    rel_br_all = mean(rel_br_mean),
    rel_br_all_sd = sd(rel_br_mean),
    .groups = "drop"
  ) %>%
  # normalize
  mutate(
    rel_amount_all = rel_amount_all / sum(rel_amount_all),
    rel_br_all = rel_br_all / sum(rel_br_all[std_brGDGT])
  )
head(lc_data_sum)

# cache
lc_data_all %>% write_rds("cache/lc_data_all.rds")
lc_data_avg %>% write_rds("cache/lc_data_avg.rds")
lc_data_sum %>% write_rds("cache/lc_data_sum.rds")
```

# Compound class abundances

```{r}
# calculate 
class_abundances_all <- 
  bind_rows(lc_data_all, gc_data_all) %>%
  group_by(sample_ID, compound_class) %>%
  summarize(
    n_peaks = n(),
    total.ug = sum(amount.ug),
    .groups = "drop"
  ) %>%
  group_by(sample_ID) %>%
  mutate(rel_amount = total.ug / sum(total.ug)) %>%
  ungroup()
head(class_abundances_all)

# averages of biological replicates
class_abundances_avg <- class_abundances_all %>%
  left_join(samples, by = "sample_ID") %>%
  group_by(exp_ID, compound_class) %>%
  summarize(
    rel_amount_mean = mean(rel_amount),
    rel_amount_sd = sd(rel_amount),
    .groups = "drop"
  )
head(class_abundances_avg)

# overall summary
class_abundances_sum <- class_abundances_avg %>%
  group_by(compound_class) %>%
  summarize(
    rel_amount_all = mean(rel_amount_mean),
    rel_amount_all_sd = sd(rel_amount_mean),
    .groups = "drop"
  ) %>%
  # normalize
  mutate(rel_amount_all = rel_amount_all / sum(rel_amount_all))
head(class_abundances_sum)

# cache
class_abundances_all %>% write_rds("cache/class_abundances_all.rds")
class_abundances_avg %>% write_rds("cache/class_abundances_avg.rds")
class_abundances_sum %>% write_rds("cache/class_abundances_sum.rds")
```

# Indices

```{r}
# derived indices
indices_all <- lc_data_all %>% 
  filter(std_brGDGT) %>%
  select(sample_ID, compound, rel_br) %>%
  mutate(compound = str_remove(compound, "brGDGT ")) %>%
  pivot_wider(values_from = rel_br, names_from = compound) %>%
  mutate(
    MBT_prime_5Me = (Ia + Ib + Ic) / (Ia + Ib + Ic + IIa + IIb + IIc + IIIa),
    CBT_5Me = - log10( (Ib + IIb) / (Ia + IIa) ),
    DC_percent = (Ib + 2 * Ic + IIb) / (Ia + Ib + Ic + IIa + IIb) * 100
  ) 
head(indices_all)

# averages of biological replicates
indices_avg <- indices_all %>%
  left_join(samples, by = "sample_ID") %>%
  group_by(exp_ID) %>%
  summarize(
    n_reps = n(),
    MBT_prime_5Me_mean = mean(MBT_prime_5Me),
    MBT_prime_5Me_sd = sd(MBT_prime_5Me),
    CBT_5Me_mean = mean(CBT_5Me),
    CBT_5Me_sd = sd(CBT_5Me),
    DC_percent_mean = mean(DC_percent),
    DC_percent_sd = sd(DC_percent),
    .groups = "drop"
  )
head(indices_avg)

# cache
indices_all %>% write_rds("cache/indices_all.rds")
indices_avg %>% write_rds("cache/indices_avg.rds")
```

# Environmental data

```{r}
# temperature types
ttypes <- tibble::tribble(
  ~sample_type, ~temperature_type,
  "Lacustrine SPM", "Water temp.",
  "Soil", "WMT (air)",
  "Lacustrine Sediment", "MAF (air)",
  "Peat", "WMT (air)",
  "Bone", "WMT (air)",
  "Marine Sediment", "SST"
)

# environmental data
env_data <- suppressWarnings(
  readxl::read_excel("data/environmental_data.xlsx", 
                     col_types = c(rep("text", 3), rep("numeric", 27))
  )) %>%
  mutate(
    CBT_5Me_infinite = is.infinite(CBT_5Me),
    CBT_5Me = ifelse(CBT_5Me_infinite, NA_real_, CBT_5Me)
  ) %>%
  left_join(ttypes, by = "sample_type") %>%
  relocate(temperature_type, .after = temperature) %>%
  relocate(CBT_5Me_infinite, .after = CBT_5Me)
head(env_data)

# cache
env_data %>% write_rds("cache/env_data.rds")
```

# Statistics

## dataset

```{r}
# culture data
culture_regs_data <- 
  indices_all %>%
  left_join(
    select(growth_rates_all, sample_ID, growth_rate.1_day), 
    by = "sample_ID"
  ) %>%
  left_join(samples, by = "sample_ID") %>%
  left_join(experiments, by = "exp_ID") %>%
  select(sample_ID, exp_ID, MBT_prime_5Me, CBT_5Me, temperature, pH, `% O2`, `growth rate` = growth_rate.1_day) %>%
  pivot_longer(c(MBT_prime_5Me, CBT_5Me), names_to = "index", values_to = "index_value") %>%
  pivot_longer(c(temperature, pH, `% O2`, `growth rate`), names_to = "variable", values_to = "var_value") %>%
  left_join(select(experiments, exp_ID, temperature, pH), by = "exp_ID") %>%
  mutate(category = "Culture")

# model fit data
data_for_model_fits <- 
  bind_rows(
    # env data
    env_data %>% 
      select(dataset = sample_type, temperature, temperature_type, pH, MBT_prime_5Me, CBT_5Me) %>%
      pivot_longer(c(MBT_prime_5Me, CBT_5Me), names_to = "index", values_to = "index_value") %>%
      pivot_longer(c(temperature, pH), names_to = "variable", values_to = "var_value") %>%
      # compare temperature with MBT and pH with CBT
      filter((index == "MBT_prime_5Me" & variable == "temperature") | (index == "CBT_5Me" & variable == "pH")) %>%
      # remove missing values, invalid comparisions, and outliers like MBT == 1
      filter(
        !is.na(index_value), !is.na(var_value), 
        !(dplyr::near(index_value, 1) & index == "MBT_prime_5Me"),
        !(dataset == "Marine Sediment" & variable == "pH")
      ) %>%
      mutate(
        category = "Environmental",
        variable = ifelse(variable == "temperature", temperature_type, variable)
      ),
    # env in situ data
    env_data %>% 
      filter(sample_type == "Soil") %>% 
      select(dataset = sample_type, index_value = MBT_prime_5Me, starts_with("in situ")) %>%
      pivot_longer(
        cols = starts_with("in situ"), names_to = "variable", values_to = "var_value"
      ) %>%
      mutate(
        index = "MBT_prime_5Me", 
        category = "Environmental",
      ) %>%
      filter(!is.na(var_value)),
    # full culture datasets
    culture_regs_data %>%
      filter(variable != "% O2") %>% mutate(dataset = "all data"),
    # O2 culturedataset
    culture_regs_data %>%
      filter(variable == "% O2", temperature == 25, pH == 5.5) %>%
      mutate(dataset = sprintf("%s\U00B0\U0043, pH %.1f", temperature, pH)),
    # pH at different temps
    culture_regs_data %>%
      filter(variable == "pH", index == "MBT_prime_5Me") %>%
      group_by(temperature) %>%
      mutate(dataset = sprintf("%s\U00B0\U0043", temperature)),
    # growth rate at different temps
    culture_regs_data %>%
      filter(variable == "growth rate", index == "MBT_prime_5Me") %>%
      group_by(temperature) %>%
      mutate(dataset = sprintf("%s\U00B0\U0043", temperature))
  )
```


## calculations

```{r}
# get linear regression fits
linear_fits <- 
  data_for_model_fits %>%
  nest(data = c(-category, -dataset, -variable, -index)) %>%
  mutate(
    n = map_int(data, nrow),
    pearson_r = map_dbl(data, ~cor(.x$index_value, .x$var_value, method = "pearson")),
    pearson_pval = map_dbl(data, ~cor.test(.x$index_value, .x$var_value, method = "pearson")$p.value),
    fit = map(data, ~lm(index_value ~ var_value, data = .x)),
    R2 = ifelse(pearson_pval < 0.05, map_dbl(fit, ~broom::glance(.x)$adj.r.squared), NA_real_),
    params = map(fit, broom::tidy)
  ) %>% 
  unnest(params) %>%
  select(-data, -fit, -statistic, -p.value) %>%
  mutate(
    term = factor(term) %>%
      fct_recode(`intercept` = "(Intercept)", `slope` = "var_value"),
    estimate = ifelse(pearson_pval < 0.05, estimate, NA_real_),
    std.error = ifelse(pearson_pval < 0.05, std.error, NA_real_)
  ) %>%
  pivot_wider(
    names_from = term,
    values_from = c(estimate, std.error), #, p.value),
    names_sep = "_"
  ) %>%
  rename(intercept = estimate_intercept, slope = estimate_slope) %>%
  select(index, category, dataset, variable, n, pearson_r, pearson_pval, matches("slope"), matches("intercept"), R2) %>%
  mutate()
```

## comparisons

```{r}
# differences to culture data (using dummy variable)
culture_comparison <- 
  inner_join(
    data_for_model_fits %>%
      filter(category == "Environmental") %>%
      nest(env_data = c(-dataset, -variable, -index)),
    data_for_model_fits %>%
      filter(
        category == "Culture",
        dataset == "all data",
        (index == "MBT_prime_5Me" & variable == "temperature") |
          (index == "CBT_5Me" & variable == "pH")
      ) %>% 
      nest(culture_data = c(-index)),
    by = "index"
  ) %>%
  mutate(
    fit = map2(env_data, culture_data,
               ~lm(index_value ~ culture + var_value + var_value * culture,
                   data = bind_rows(
                     mutate(.x, culture = 0), 
                     mutate(.y, culture = 1) 
    ))),
    # Note: could use chow test here as well but provides less info
    # chow_pval = map2(data, culture_data,
    #   ~{print(nrow(.x)); print(nrow(.y)); safe_chow(.x$index, as.matrix(.x$value, ncol = 1), 
    #                  .y$index, as.matrix(.y$value, ncol = 1))}
    # ),
    params = map(fit, broom::tidy)
  ) %>% unnest(params) %>% 
  select(-env_data, -culture_data, -fit) %>%
  filter(str_detect(term, "culture")) %>%
  mutate(
    term = factor(term) %>% 
      fct_recode(`diff_intercept` = "culture", `diff_slope` = "culture:var_value")) %>%
  select(-statistic) %>%
  pivot_wider(names_from = term, values_from = c(estimate, std.error, p.value), names_sep = "_") %>%
  rename(diff_intercept = estimate_diff_intercept,  diff_slope = estimate_diff_slope) %>%
  select(dataset, index, variable, matches("slope"), matches("intercept")) %>%
  filter(!is.na(diff_slope))

# combination
regression_data <- 
  linear_fits %>% 
  left_join(culture_comparison, by = c("dataset", "index", "variable")) %>%
  mutate(order1 = ifelse(category == "Culture", variable, dataset),
         order2 = ifelse(category == "Culture", dataset, variable)) %>%
  arrange(desc(index), category, order1, desc(order2)) %>%
  select(-order1, -order2)
head(regression_data)

# caching
regression_data %>% write_rds("cache/regression_data.rds")
```

# Mass spectra

```{r}
# peaks
peaks <- 
  readxl::read_excel("data/mass_spectra_peaks.xlsx") %>%
  mutate(
    peak_join_mass = sprintf("%.1f", mass),
    compound = as_factor(compound)
  )

# spectra
spectra <-
  # get unique compounds from peaks list
  peaks %>% select(compound) %>% unique() %>%
  # read in spectra
  mutate(
    data = map(compound, ~readxl::read_excel("data/mass_spectra.xlsx", sheet = as.character(.x), skip = 7))
  ) %>%
  unnest(data) %>%
  # normalize
  group_by(compound) %>%
  mutate(rel_intensity = Intensity / max(Intensity, na.rm = TRUE)) %>%
  # find peaks
  mutate(signal = rel_intensity > 1e-4, peak_nr = cumsum(c(0, diff(signal)) > 0)) %>% 
  # identify peak apex intensity
  group_by(compound, peak_nr) %>%
  mutate(
    peak_rel_intensity = ifelse(peak_nr > 0, max(rel_intensity), NA_real_),
    peak_mass = Mass[rel_intensity == peak_rel_intensity[1]][1],
    peak_join_mass = sprintf("%.1f", peak_mass)
  ) %>%
  ungroup() %>%
  # add peaks in
  left_join(peaks, by = c("compound", "peak_join_mass")) %>%
  # figure out peak boundaries
  group_by(compound) %>%
  mutate(
    peak_id = ifelse(!signal, "no peak", peak_join_mass) %>% as_factor(),
    peak_grp = c(0, cumsum(diff(as.integer(peak_id)))),
    start_peak = c(0, diff(peak_grp)) != 0,
    end_peak = c(diff(peak_grp), 0) != 0
  ) %>% 
  # add in peak boundary start/end points
  {
    bind_rows(
      ., 
      filter(., start_peak | end_peak) %>% mutate(rel_intensity = 0, start_peak = FALSE, end_peak = FALSE)
    )
  } %>% 
  ungroup()
  # make compound into factor
  
head(spectra)

# caching
spectra %>% write_rds("cache/spectra.rds")
```

# Chromatograms

```{r}
# chroms peaks
chroms_peaks <- readxl::read_excel("data/chroms_peaks.xlsx")

# retention time alignment
calib <- chroms_peaks %>%
  pivot_wider(names_from = sample, values_from = peak_rt) %>%
  filter(!is.na(soil), !is.na(susi)) %>%
  with(lm(susi ~ soil))

# chromatograms
chroms <- 
  tibble(
    sheet = c("susi_1050-brGDGTIIIa", "susi_1048-brGDGTIIIb", "soil_1050", "soil_1048"),
    mass = str_extract(sheet, "[0-9]+"),
    sample = str_extract(sheet, "^[^_]+"),
    data = map(sheet, ~readxl::read_excel("data/chroms.xlsx", sheet = .x, skip = 3))
  ) %>% 
  unnest(data) %>%
  mutate(time_join = sprintf("%.1f", Time)) %>%
  left_join(
    mutate(chroms_peaks, time_join = sprintf("%.1f", peak_rt)), 
    by = c("sample", "time_join")
  ) %>%
  { mutate(., time_shifted = predict(calib, newdata = select(., soil = Time))) } %>%
  mutate(time = ifelse(sample == "soil", time_shifted, Time)) %>%
  group_by(sheet, compound, mass) %>%
  mutate(
    peak_intensity = ifelse(!is.na(compound), max(Intensity), NA_real_),
    peak_rt = ifelse(!is.na(compound), time[Intensity == peak_intensity][1], NA_real_)
  ) %>%
  ungroup()

head(chroms)

# caching
chroms %>% write_rds("cache/chroms.rds")
```
