==> 02_tables.Rmd <==
---
title: "Tables and Datasets"
date: "`r format(Sys.Date(), '%d %b %Y')`"
output:
  html_document: 
    df_print: paged
    number_sections: no
    toc: yes
    toc_float: true
    toc_depth: 3
    code_folding: show
editor_options:
  chunk_output_type: console
---

```{r setup, echo = FALSE, message=FALSE, warning=FALSE}
# load libraries
library(tidyverse) 
library(readxl)

# source all relevant scripting files
source(file.path("scripts", "table_functions.R"))

# global knitting options
knitr::opts_chunk$set(echo = TRUE)
```

> NOTE: please knit the `01_calculations.Rmd` notebook first to prepare the data used in this notebook.

# Load experiment data

```{r}
# load experiments and samples
experiments <- read_rds("cache/experiments.rds")
samples <- read_rds("cache/samples.rds")
```

# Table 1: statistics

```{r}
# p-value formatting
format_pval <- function(x) {
  case_when(
    is.na(x) ~ NA_character_,
    x < 1e-3 ~ "<0.001 (***)",
    x < 1e-2 ~ "<0.01 (**)",
    x < 5e-2 ~ "<0.05 (*)",
    TRUE ~ sprintf("%.2f (\U2014)", x)
  )
}

# table 1
table_1 <- read_rds("cache/regression_data.rds") %>%
  mutate(
    index = as_factor(index) %>% fct_recode("MBT'5Me" = "MBT_prime_5Me", "CBT5Me" = "CBT_5Me"),
    pearson_r = sprintf("%+.2f", pearson_r),
    pearson_pval = format_pval(pearson_pval),
    # slope and intercept
    slope_decimals = ifelse(index == "MBT'5Me", 3, 2),
    slope = ifelse(
      !is.na(slope),
      sprintf("%%.%.0ff\U00B1%%.%.0ff", slope_decimals, slope_decimals) %>%
        sprintf(slope, std.error_slope),
      NA_character_
    ),
    intercept_decimals = ifelse(index == "MBT'5Me", 2, 1),
    intercept = ifelse(
      !is.na(intercept), 
      sprintf("%%.%.0ff\U00B1%%.%.0ff", intercept_decimals, intercept_decimals) %>%
        sprintf(intercept, std.error_intercept),
      NA_character_
    ),
    R2 = ifelse(!is.na(R2), sprintf("%.2f", R2), NA_character_),
    # diff slope and intercept
    diff_slope =
      case_when(
        !is.na(diff_slope) ~
          sprintf("%%.%.0ff\U00B1%%.%.0ff", slope_decimals, slope_decimals) %>%
              sprintf(diff_slope, std.error_diff_slope),
        TRUE ~ NA_character_),
    diff_slope_sig = format_pval(p.value_diff_slope),
    diff_intercept =
      case_when(
        !is.na(diff_intercept) ~
          sprintf("%%.%.0ff\U00B1%%.%.0ff", intercept_decimals, intercept_decimals) %>%
              sprintf(diff_intercept, std.error_diff_intercept),
        TRUE ~ NA_character_),
    diff_intercept_sig = format_pval(p.value_diff_intercept)
  ) %>%
  select(index, category, dataset, variable, pearson_r, pearson_pval, 
         slope, intercept, R2, diff_slope, diff_slope_sig, diff_intercept, diff_intercept_sig)

table_1 %>% export_to_excel(file = "tables/table_1_regressions_data.xlsx")
table_1
```


# Table S1: growth rates

```{r}
# load all data (combine growth rates and generations)
table_growth_rates_all <- read_rds("cache/growth_rates_all.rds") %>%
  left_join(
    read_rds("cache/generations_all.rds"),
    by = "sample_ID"
  ) %>%
  left_join(samples, by = "sample_ID") %>%
  left_join(experiments, by = "exp_ID") %>%
  select(sample_ID, temperature, pH, `% O2`, rep, 
         growth_rate.1_day, growth_rate_se.1_day, 
         n_generations) %>%
  arrange(desc(`% O2`), temperature, pH, rep)
  
# load averages of biological replicates
table_growth_rates_avg <- read_rds("cache/growth_rates_avg.rds") %>%
   left_join(
    read_rds("cache/generations_avg.rds") %>% select(-n_reps),
    by = "exp_ID"
  ) %>%
  left_join(experiments, by = "exp_ID") %>%
  select(exp_ID, temperature, pH, `% O2`, n_reps, 
         growth_rate_mean.1_day, growth_rate_sd.1_day, 
         n_generations_mean, n_generations_sd) %>%
  arrange(desc(`% O2`), temperature, pH)

# dataset S1
export_to_excel(
  growth_rates_all = table_growth_rates_all %>% select(-sample_ID),
  growth_rates = table_growth_rates_avg %>% select(-exp_ID),
  file = "tables/dataset_S1-growth_rates.xlsx"
)

# table S1
table_S1 <- 
  bind_rows(
    table_growth_rates_all %>% 
      mutate(
        rep = as.character(rep),
        value = sprintf("%.2f\U00B1%.2f", growth_rate.1_day, growth_rate_se.1_day)
      ),
    table_growth_rates_avg %>% 
      mutate(
        rep = "average", 
        value = sprintf("%.2f\U00B1%.2f", growth_rate_mean.1_day, growth_rate_sd.1_day)
      ),
    table_growth_rates_avg %>%
      mutate(
        rep = "n_gens", 
        value = sprintf("%.2f\U00B1%.2f", n_generations_mean, n_generations_sd)
      )
  ) %>%
  select(temperature, pH, `% O2`, rep, value) %>%
  pivot_wider(names_from = "rep", values_from = value) %>%
  arrange(desc(`% O2`), temperature, pH)

table_S1 %>% export_to_excel(file = "tables/table_S1_growth_rates.xlsx")
table_S1
```

# Table S2: membrane composition

```{r}
# load all data
table_class_abundances_all <- read_rds("cache/class_abundances_all.rds") %>%
  left_join(samples, by = "sample_ID") %>%
  left_join(experiments, by = "exp_ID") %>%
  select(sample_ID, temperature, pH, `% O2`, rep, rel_amount) %>%
  arrange(desc(`% O2`), temperature, pH, rep)

# load averages of biological replicates
table_class_abundances_avg <- read_rds("cache/class_abundances_avg.rds") %>%
  left_join(experiments, by = "exp_ID") %>%
  select(exp_ID, temperature, pH, `% O2`, compound_class, rel_amount_mean, rel_amount_sd) %>%
  arrange(desc(`% O2`), temperature, pH, compound_class) 

# load overall summary
table_class_abundances_sum <- read_rds("cache/class_abundances_sum.rds")
 
# dataset S1
export_to_excel(
  composition_data_all = table_class_abundances_all %>% select(-sample_ID),
  composition_data = table_class_abundances_avg %>% select(-exp_ID),
  composition_data_summary = table_class_abundances_sum,
  file = "tables/dataset_S1-membrane_composition_data.xlsx"
)

# SI table
table_S2 <- 
  bind_rows(
    table_class_abundances_avg %>% 
      mutate(info = sprintf("%s\U00B0\U0043\npH %s\n%s%% O2", temperature, pH, `% O2`)) %>%
      rename(amount = rel_amount_mean, err = rel_amount_sd),
    table_class_abundances_sum %>% 
      rename(amount = rel_amount_all, err = rel_amount_all_sd) %>%
      mutate(info = "All")
  ) %>%
  mutate(value = sprintf("%.1f\n\U00B1%.1f", 100*amount, 100*err)) %>%
  select(info, compound_class, value) %>%
  pivot_wider(names_from = info, values_from = value) %>%
  arrange(compound_class)
table_S2 %>% export_to_excel(file = "tables/table_S2_membrane_composition_data.xlsx")
table_S2
```


# Table S3: GC data

```{r}
# load all data
table_gc_data_all <- read_rds("cache/gc_data_all.rds") %>%
  left_join(samples, by = "sample_ID") %>%
  left_join(experiments, by = "exp_ID") %>%
  select(sample_ID, temperature, pH, `% O2`, rep, compound, compound_class, rel_amount) %>%
  arrange(desc(`% O2`), temperature, pH, rep) %>%
  filter(rel_amount > 0)

# load averages of biological replicates
table_gc_data_avg <- read_rds("cache/gc_data_avg.rds") %>%
  left_join(experiments, by = "exp_ID") %>%
  select(exp_ID, temperature, pH, `% O2`, compound, n_reps, rel_amount_mean, rel_amount_sd) %>%
  arrange(desc(`% O2`), temperature, pH, compound) 

# load overall summary
table_gc_data_sum <- read_rds("cache/gc_data_sum.rds")

# dataset export
export_to_excel(
  GC_data_all = table_gc_data_all %>% select(-sample_ID),
  GC_data = table_gc_data_avg %>% select(-exp_ID),
  GC_data_summary = table_gc_data_sum,
  file = "tables/dataset_S1-GC_data.xlsx"
)

# SI table
table_S3 <- 
  bind_rows(
    table_gc_data_avg %>% 
      mutate(info = sprintf("%s\U00B0\U0043\npH %s\n%s%% O2", temperature, pH, `% O2`)) %>%
      rename(amount = rel_amount_mean, err = rel_amount_sd),
    table_gc_data_sum %>% 
      rename(amount = rel_amount_all, err = rel_amount_all_sd) %>%
      mutate(info = "All")
  ) %>%
  mutate(
    value = ifelse(amount < 0.01, "< 1\n", sprintf("%.1f\n\U00B1%.1f", 100*amount, 100*err))
  ) %>%
  select(info, compound, value) %>%
  pivot_wider(names_from = info, values_from = value) %>%
  arrange(compound)
table_S3 %>% export_to_excel(file = "tables/table_S3_GC_data.xlsx")
table_S3
```

# Table S4: LC data

```{r}
# load all data
table_lc_data_all <- read_rds("cache/lc_data_all.rds") %>%
  left_join(samples, by = "sample_ID") %>%
  left_join(experiments, by = "exp_ID") %>%
  select(sample_ID, temperature, pH, `% O2`, rep, compound, compound_class, 
         rel_amount, rel_to_standard_brGDGTs = rel_br) %>%
  arrange(desc(`% O2`), temperature, pH, rep)

# load averages of biological replicates
table_lc_data_avg <- read_rds("cache/lc_data_avg.rds") %>%
  left_join(experiments, by = "exp_ID") %>%
  select(exp_ID, temperature, pH, `% O2`, compound, n_reps, 
         rel_amount_mean, rel_amount_sd, 
         rel_to_standard_brGDGTs_mean = rel_br_mean,
         rel_to_standard_brGDGTs_sd = rel_br_sd) %>%
  arrange(desc(`% O2`), temperature, pH, compound) 

# load overall summary
table_lc_data_sum <- read_rds("cache/lc_data_sum.rds") %>%
   rename(
     is_standard_brGDGT = std_brGDGT,
     rel_to_standard_brGDGTs_all = rel_br_all, 
     rel_to_standard_brGDGTs_all_sd = rel_br_all_sd
  )

# dataset export
export_to_excel(
  LC_data_all = table_lc_data_all %>% select(-sample_ID),
  LC_data = table_lc_data_avg %>% select(-exp_ID),
  LC_data_summary = table_lc_data_sum,
  file = "tables/dataset_S1-LC_data.xlsx"
)

# SI table
table_S4 <- 
  bind_rows(
    table_lc_data_avg %>% 
      mutate(info = sprintf("%s\U00B0\U0043\npH %s\n%s%% O2", temperature, pH, `% O2`)) %>%
      rename(amount = rel_amount_mean, err = rel_amount_sd),
    table_lc_data_sum %>% 
      rename(amount = rel_amount_all, err = rel_amount_all_sd) %>%
      mutate(info = "All")
  ) %>%
  mutate(
    decimals = find_signif_decimals(100*amount, 2) %>% { ifelse(. < 2, ., 2)},
    value = 
      ifelse(
        dplyr::near(amount, 0), "n.q.",
        sprintf("%%.%.0ff\n\U00B1%%.%.0ff", decimals, decimals) %>%
          sprintf(100*amount, 100*err)
      )
  ) %>%
  select(info, compound, value) %>%
  pivot_wider(names_from = info, values_from = value) %>%
  arrange(compound)
table_S4 %>% export_to_excel(file = "tables/table_S4_LC_data.xlsx")
table_S4
```

# Table S5: derived indices and %br

```{r}
# derived indices
table_indices_all <- read_rds("cache/indices_all.rds") %>%
  left_join(samples, by = "sample_ID") %>%
  left_join(experiments, by = "exp_ID") %>%
  select(sample_ID, temperature, pH, `% O2`, rep, 
         MBT_prime_5Me, CBT_5Me, DC_percent) %>%
  arrange(desc(`% O2`), temperature, pH, rep)

# load averages of biological replicates
table_indices_avg <- read_rds("cache/indices_avg.rds") %>%
  left_join(experiments, by = "exp_ID") %>%
  select(exp_ID, temperature, pH, `% O2`, MBT_prime_5Me_mean:DC_percent_sd) %>%
  arrange(desc(`% O2`), temperature, pH) 

# dataset export
export_to_excel(
  index_data_all = table_indices_all %>% select(-sample_ID),
  index_data = table_indices_avg %>% select(-exp_ID),
  file = "tables/dataset_S1-index_data.xlsx"
)

# SI table
table_S5 <- 
  bind_rows(
    left_join(
      table_indices_avg %>% select(-ends_with("sd")) %>% 
        pivot_longer(cols = ends_with("mean")) %>% 
        mutate(name = str_remove(name, "_mean")),
      table_indices_avg %>% select(-ends_with("mean")) %>% 
        pivot_longer(cols = ends_with("sd")) %>% 
        mutate(name = str_remove(name, "_sd")) %>% select(exp_ID, name, sd = value),
      by = c("exp_ID", "name")
    ) %>%
      transmute(
        name = name,
        info = sprintf("%s\U00B0\U0043\npH %s\n%s%% O2", temperature, pH, `% O2`),
        decimals = find_signif_decimals(value, 2) %>% { ifelse(. < 2, ., 2)},
        value = sprintf("%%.%.0ff\n\U00B1%%.%.0ff", decimals, decimals) %>%
          sprintf(value, sd)
      ),
    read_rds("cache/lc_data_avg.rds") %>%
      left_join(experiments, by = "exp_ID") %>% 
      filter(std_brGDGT) %>%
      transmute(
        name = compound,
        info = sprintf("%s\U00B0\U0043\npH %s\n%s%% O2", temperature, pH, `% O2`),
        decimals = if_else(rel_br_mean > 0.1, 1, 2), 
        value = 
          ifelse(
            dplyr::near(rel_br_mean, 0), "n.q.",
            sprintf("%%.%.0ff\n\U00B1%%.%.0ff", decimals, decimals) %>%
              sprintf(100*rel_br_mean, 100*rel_br_sd)
          )
      )
  ) %>%
  select(info, name, value) %>%
  pivot_wider(names_from = info, values_from = value)
table_S5 %>% export_to_excel(file = "tables/table_S5_indices_plus_perc_br.xlsx")
table_S5
```


# Dataset S1: growth rate and lipid data

```{r}
export_to_excel(
  # renaming tabs for clarity
  growth_rates = table_growth_rates_all %>% select(-sample_ID),
  growth_rates_summary = table_growth_rates_avg %>% select(-exp_ID),
  composition_data = table_class_abundances_all %>% select(-sample_ID),
  composition_data_summary = table_class_abundances_avg %>% select(-exp_ID),
  LC_data = table_lc_data_all %>% select(-sample_ID),
  LC_data_summary = table_lc_data_avg %>% select(-exp_ID),
  GC_data = table_gc_data_all %>% select(-sample_ID),
  GC_data_summary = table_gc_data_avg %>% select(-exp_ID),
  index_data = table_indices_all %>% select(-sample_ID),
  index_data_summary = table_indices_avg %>% select(-exp_ID),
  file = "tables/dataset_S1.xlsx"
)
```


# Dataset S2: environmental / culture climate proxy data

```{r}
# combine env and culture data
env_w_culture_data <- 
  bind_rows(
    # env data
    read_rds("cache/env_data.rds"),
    # culture data
    read_rds("cache/indices_all.rds") %>%
      left_join(samples, by = "sample_ID") %>%
      left_join(experiments, by = "exp_ID") %>%
      arrange(desc(`% O2`), temperature, pH, rep) %>%
      mutate(
        reference = "this study",
        sample_id = sprintf("%.0f\U00B0\U0043, pH %.1f, %.0f %% O2, replicate %d",
                            temperature, pH, `% O2`, rep),
        sample_type = "Culture",
        temperature_type = "in situ temp.",
        CBT_5Me_infinite = FALSE,
        DC = DC_percent / 100
      ) %>%
      select(reference, sample_id, sample_type, temperature, temperature_type, 
             MBT_prime_5Me, CBT_5Me, CBT_5Me_infinite, DC, Ia:IIIc)
  ) 

export_to_excel(
  dataset = env_w_culture_data,
  regressions = read_rds("cache/regression_data.rds"),
  file = "tables/dataset_S2.xlsx"
)
```
