==> 04_analysis_low_O2.Rmd <==
---
title: "Calculations"
date: "`r format(Sys.Date(), '%d %b %Y')`"
output:
  html_document: 
    df_print: paged
    number_sections: no
    toc: yes
    toc_float: true
    toc_depth: 3
    code_folding: show
editor_options:
  chunk_output_type: console
---

```{r setup, echo = FALSE, message=FALSE, warning=FALSE}
# load libraries
library(tidyverse) 
library(readxl)

# source all relevant scripting files
source("scripts/table_functions.R")
source("scripts/growth_functions.R")

# global knitting options
knitr::opts_chunk$set(echo = TRUE)
```

# Experiment data

```{r}
# OD multipliers
ministat_OD_scaling <- 3.96 # as per OD measurements culture tube vs. 

# load experiments
experiments <- read_xlsx("data/experiments_low_O2.xlsx") |>
  mutate( 
    # temp and %O2 are always whole integers in this study
    temperature = as.integer(temperature),
    `% O2` = as.integer(o2),
    # OD multiplier depends on ministat use
    OD_multiplier = if_else(!is.na(ministat) & ministat, 1/ministat_OD_scaling, 1)
  )

# cache
experiments |> write_rds("cache/experiments_low_O2.rds")
```

# Normalized OD

```{r}
# load growth curve data
growth_data_all <- 
  read_excel("data/growth_data_low_O2.xlsx") |>
  left_join(
    experiments |> select(exp_ID, OD_multiplier),
    by = "exp_ID"
  ) |>
  mutate(
    growth_phase = case_when(
      death_phase ~ "death",
      extended_lag ~ "long lag",
      TRUE ~ "growth"
    ),
    OD_norm = OD_multiplier * OD
  )
head(growth_data_all)

# cache
growth_data_all |> write_rds("cache/growth_data_all_low_O2.rds")
```

# Growth rates

```{r}
# growth data
growth_data <- 
  growth_data_all |>
  # experiments$exp_ID |> unique() |> cat(sep = "\", \"")
  # works
  #filter(exp_ID %in% c("exp65", "exp66", "exp67", "exp69", "exp71")) 
  #filter(exp_ID %in% c("exp72", "exp75", "exp76", "exp77", "exp79", "exp80"))
  filter(exp_ID %in% c("exp65", "exp66", "exp67", "exp69", "exp71", "exp72", "exp75", "exp76", "exp77", "exp79", "exp80"))
  #filter(exp_ID %in% c("exp79"))
  # ministat doesn't work (none of them do)
  #filter(exp_ID %in% c("exp62"))
  #filter(exp_ID %in% c("exp67")) 
  # batch doesn't work
  #filter(exp_ID %in% c("exp70", "exp73", "exp74", "exp78")) 
  
# calculate growth rates
growth_rates_all <- 
  growth_data |>
  # fit exponential equation to growth curves
  estimate_growth_curve_parameters(
    time = time.hours, N = OD,
    group_by = c(exp_ID, replicate),
    type = "exponential",
    window_width_n = 3, # 5 data point window
    find_max_r_filter = N > 0.1 * max(N) # max rate at later points
  ) |>
  # calculate growth rates in 1/day
  mutate(
    growth_rate.1_day = r * 24, 
    growth_rate_se.1_day = r_se * 24,
    .after = "replicate"
  )
head(growth_rates_all)

# calculate averages of biological replicates
growth_rates_avg <- 
  growth_rates_all |>
  filter(!is.na(r)) |>
  summarize(
    .by = "exp_ID",
    n_reps = n(),
    growth_rate_mean.1_day = mean(growth_rate.1_day), 
    growth_rate_sd.1_day = sd(growth_rate.1_day)
  )
head(growth_rates_avg)

# cache
growth_rates_all |> write_rds("cache/growth_rates_all_low_O2.rds")
growth_rates_avg |> write_rds("cache/growth_rates_avg_low_O2.rds")
######```

# Growh curves

######```{r "figure_S1_growth_curves", fig.width=10, fig.height=10, warning=FALSE}
# plotting data
plot_growth_data <- 
  #read_rds("cache/growth_data_all_low_O2.rds") |>
  growth_data |> 
  #x |>
  ungroup() |>
  select(exp_ID, replicate, time.hours, OD = OD) |>
  nest(OD_data = c(time.hours, OD)) |>
  left_join(growth_rates_all, by = c("exp_ID", "replicate")) |>
  left_join(experiments, by = "exp_ID") |>
  mutate(
    pH_info = sprintf("pH %s", pH) |> as.factor() |> fct_rev(),
    temp_info = sprintf("%s\U00B0\U0043", temperature), 
    O2_info = sprintf("%s%% O2", paste(`% O2`)) |> factor(levels = c("1% O2", "5% O2", "21% O2"))
  ) 

# plot
plot_growth_data |>
  ggplot() + 
  aes(x = time.hours/24, y = OD, group = paste(exp_ID, replicate), 
      color = factor(replicate) #O2_info,
    ) +
  geom_point(data = function(df) unnest(df, OD_data), map = aes(shape = exp_ID), size = 3, alpha = 0.5) +
  geom_line(data = function(df)
    df |> filter(is.na(error)) |>
    generate_exponential_curve(
      time = time.hours, N = OD, blank = blank,
      time_min = lag_time, time_max = mid_exp_time, # N_max = 0.5
    ),
    linewidth = 1) +
  #geom_hline(yintercept = OD_cutoff, linetype = 2) +
  theme_figure(text_size = 14) +
  theme(
    legend.position = "bottom", legend.direction = "horizontal", 
    legend.text = element_text(size = 20)
  ) +
  scale_color_manual(values = pH_colors) +
  labs(x = "Time [days]", y = "Optical Density", color = NULL) +
  #guides(shape = guide_none()) +
  #facet_grid(pH_info ~ temp_info, scales = "free") 
  #facet_wrap(~exp_ID, scales = "free")
  facet_wrap(~replicate, scales = "free_x")
```
